move phaseItemMasonry to ts to see if it can find the bug
