+++ src/types/animations.ts
@@ -1,0 +1,16 @@
+export interface Point {
+  top: number
+  left: number
+}
+
+export interface Dims {
+  height: number
+  width: number
+}
+
+export interface Coords {
+  x: number
+  y: number
+}
+
+export interface BBox extends Point, Dims {}
