+++ src/universal/components/ReflectionGroup/ReflectionGroup.tsx
@@ -1,18 +1,14 @@
-import {DropTarget as dropTarget} from '@mattkrick/react-dnd'
+import {ConnectDropTarget} from '@mattkrick/react-dnd'
 import {ReflectionGroup_meeting} from '__generated__/ReflectionGroup_meeting.graphql'
 import {ReflectionGroup_reflectionGroup} from '__generated__/ReflectionGroup_reflectionGroup.graphql'
-import React, {Component, ReactElement} from 'react'
+import React, {Component} from 'react'
 import styled, {css} from 'react-emotion'
-import {commitLocalUpdate, createFragmentContainer, graphql} from 'react-relay'
+import {commitLocalUpdate} from 'react-relay'
 import DraggableReflectionCard from 'universal/components/ReflectionCard/DraggableReflectionCard'
-import Modal from 'universal/components/Modal'
-import {MasonryChildrenCache, MasonryParentCache} from 'universal/components/ReflectionCardInFlight'
 import ReflectionGroupHeader from 'universal/components/ReflectionGroupHeader'
-import withAtmosphere, {
-  WithAtmosphereProps
-} from 'universal/decorators/withAtmosphere/withAtmosphere'
+import {WithAtmosphereProps} from 'universal/decorators/withAtmosphere/withAtmosphere'
 import {STANDARD_CURVE} from 'universal/styles/animation'
-import {GROUP, REFLECTION_CARD, REFLECTION_GROUP, VOTE} from 'universal/utils/constants'
+import {GROUP, REFLECTION_CARD, VOTE} from 'universal/utils/constants'
 import getScaledModalBackground from 'universal/utils/multiplayerMasonry/getScaledModalBackground'
 import initializeModalGrid from 'universal/utils/multiplayerMasonry/initializeModalGrid'
 import {
@@ -23,22 +19,31 @@ import {
   MODAL_PADDING
 } from 'universal/utils/multiplayerMasonry/masonryConstants'
 import updateReflectionsInModal from 'universal/utils/multiplayerMasonry/updateReflectionsInModal'
-import withMutationProps, {WithMutationProps} from 'universal/utils/relay/withMutationProps'
+import {WithMutationProps} from 'universal/utils/relay/withMutationProps'
+import {
+  MasonryChildrenCache,
+  MasonryItemCache,
+  MasonryParentCache,
+  SetChildRef,
+  SetItemRef
+} from '../PhaseItemMasonry'
 
-interface Props extends WithAtmosphereProps, WithMutationProps {
-  canDrop: boolean
-  itemCache: any
-  childrenCache: MasonryChildrenCache
-  parentCache: MasonryParentCache
-  connectDropTarget: (reactEl: ReactElement<{}>) => ReactElement<{}>
+interface PassedProps {
   meeting: ReflectionGroup_meeting
   reflectionGroup: ReflectionGroup_reflectionGroup
-  retroPhaseItemId: string
-
-  setItemRef(c: HTMLElement): void
-
-  setChildRef: (groupId: string, reflectionId: string) => (c: HTMLElement) => void
+  itemCache: MasonryItemCache
+  childrenCache: MasonryChildrenCache
+  parentCache: MasonryParentCache
+  setItemRef: SetItemRef
+  setChildRef: SetChildRef
 }
+
+interface CollectedProps {
+  connectDropTarget: ConnectDropTarget
+  canDrop: boolean
+}
+
+interface Props extends WithAtmosphereProps, WithMutationProps, PassedProps, CollectedProps {}
 
 const reflectionsStyle = (
   canDrop: boolean | undefined,
@@ -134,6 +139,7 @@ class ReflectionGroup extends Component<Props> {
   componentWillUnmount() {
     this.closeGroupModal(true)
   }
+
   closeGroupModal = (isForce?: boolean) => {
     const {
       atmosphere,
@@ -142,21 +148,23 @@ class ReflectionGroup extends Component<Props> {
       itemCache,
       reflectionGroup: {isExpanded, reflectionGroupId, reflections}
     } = this.props
-    if (!isExpanded || !this.modalRef || !this.backgroundRef) return
+    if (!isExpanded || !this.modalRef || !this.backgroundRef || !parentCache.boundingBox) return
     const {
       boundingBox: {left: parentLeft, top: parentTop}
     } = parentCache
     const {style: modalStyle} = this.modalRef
     const {style: backgroundStyle} = this.backgroundRef
-    const firstItemHeight = itemCache[reflections[0].id].boundingBox.height
+    const cachedFirstItem = itemCache[reflections[0].id]
+    if (!cachedFirstItem.boundingBox) return
+    const firstItemHeight = cachedFirstItem.boundingBox.height
     const childDuration = EXIT_DURATION + MIN_VAR_ITEM_DELAY * (reflections.length - 1)
 
     // set starting item styles
     reflections.forEach((reflection, idx) => {
       const cachedItem = itemCache[reflection.id]
-      const {
-        modalEl: {style: itemStyle}
-      } = cachedItem
+      const {modalEl} = cachedItem
+      if (!modalEl) return
+      const {style: itemStyle} = modalEl
       const cardStackOffset = idx === reflections.length - 1 ? 0 : 6
       const delay = MIN_VAR_ITEM_DELAY * (reflections.length - 1 - idx)
       itemStyle.height = `${firstItemHeight}px`
@@ -325,32 +333,33 @@ class ReflectionGroup extends Component<Props> {
   }
 }
 
-const reflectionDropSpec = {
-  canDrop(props: Props, monitor) {
+const reflectionDropSpec: DropTargetSpec<Props, {}, ReflectionGroup> = {
+  canDrop(props, monitor) {
     return (
       monitor.isOver() &&
       monitor.getItem().reflectionGroupId !== props.reflectionGroup.reflectionGroupId
     )
   },
 
-  drop(props: Props, monitor) {
+  drop(props, monitor) {
     if (monitor.didDrop()) return undefined
     const {reflectionGroup} = props
     const {reflectionGroupId: targetReflectionGroupId} = reflectionGroup
-    return {dropTargetType: REFLECTION_GROUP, dropTargetId: targetReflectionGroupId}
+    return {
+      dropTargetType: DragReflectionDropTargetTypeEnum.REFLECTION_GROUP,
+      dropTargetId: targetReflectionGroupId
+    }
   }
 }
 
-const reflectionDropCollect = (connect, monitor) => ({
+const reflectionDropCollect = (connect: DropTargetConnector, monitor: DropTargetMonitor) => ({
   connectDropTarget: connect.dropTarget(),
   canDrop: monitor.canDrop()
 })
 
-export default createFragmentContainer(
-  withAtmosphere(
-    withMutationProps(
-      dropTarget(REFLECTION_CARD, reflectionDropSpec, reflectionDropCollect)(ReflectionGroup)
-    )
+export default createFragmentContainer<PassedProps>(
+  (dropTarget as any)(REFLECTION_CARD, reflectionDropSpec, reflectionDropCollect)(
+    withAtmosphere(withMutationProps(ReflectionGroup))
   ),
   graphql`
     fragment ReflectionGroup_meeting on RetrospectiveMeeting {
